+++ lib/blog_overview/widgets/widgets.dart
@@ -1,1 +1,1 @@
-export 'header.dart';
+export 'blog_overview_header.dart';
