+++ test/blog_overview/view/blog_overview_page_test.dart
@@ -7,7 +7,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:personal_blog_flutter/blog_overview/bloc/blog_overview_bloc.dart';
 import 'package:personal_blog_flutter/blog_overview/view/blog_overview_page.dart';
-import 'package:personal_blog_flutter/blog_overview/widgets/header.dart';
+import 'package:personal_blog_flutter/blog_overview/widgets/blog_overview_header.dart';
 import 'package:personal_blog_flutter/l10n/l10n.dart';
 
 import '../../helpers/helpers.dart';
@@ -77,8 +77,7 @@ void main() {
         expect(find.text(failureMessage), findsOneWidget);
       });
 
-      testWidgets('renders header with correct text and blog cards',
-          (tester) async {
+      testWidgets('renders header and blog cards', (tester) async {
         final posts = List.generate(
           3,
           (index) => BlogPost(
@@ -102,17 +101,7 @@ void main() {
             child: const BlogOverview(),
           ),
         );
-        expect(find.byType(Header), findsOneWidget);
-        final context = tester.element(find.byType(Header));
-        expect(
-          find.text(context.l10n.blogOverviewHeader.toUpperCase()),
-          findsOneWidget,
-        );
-        expect(
-          find.text(context.l10n.blogOverviewHeaderSubtitle),
-          findsOneWidget,
-        );
-
+        expect(find.byType(BlogOverviewHeader), findsOneWidget);
         expect(find.byType(BlogCard), findsNWidgets(3));
       });
     });
